+++ application/models/template_model.php
@@ -41,6 +41,8 @@ class Template_Model extends My_Model
 				}
 			}
 		}
+		
+		$this->check_for_site_updates();
 	}
 	
 	function check_for_updates()
@@ -135,6 +137,89 @@ class Template_Model extends My_Model
 			
 			//delete any variables that have been removed
 			foreach ($this->template_variables->all() as $tv) {
+				if (!in_array($tv->name, $variables) && !$tv->template_variable_id) {
+					$tv->destroy();
+				}
+			}
+		}
+	}
+	
+	function check_for_site_updates()
+	{
+		//get config
+		$config = FCPATH.'assets/site/templates/site.config.php';
+		if (!file_exists($config)) return false;
+		
+		include($config);
+		
+		if (isset($site))
+		{
+			//check for variables
+			$variables = array();
+			if (isset($site['variables']) && count($site['variables']))
+			{
+				//check each variable to see if it exists already
+				foreach ($site['variables'] as $variable)
+				{
+					if (!$this->template_variable_model->exists(array('name' => $variable['name'], 'template_variable_id' => null, 'template_id' => 0)))
+					{
+						//create variable
+						$var = $this->template_variable_model->create(array(
+							'template_id' => 0,
+							'type'  => isset($variable['type']) ? $variable['type'] : 'string',
+							'name'  => $variable['name'],
+							'label' => isset($variable['label']) ? $variable['label'] : $variable['name'],
+							'value' => isset($variable['default']) ? $variable['default'] : '',
+							'options' => isset($variable['options']) ? serialize($variable['options']) : ''
+						));
+						
+						if ($variable['type'] == 'array' && $variable['variables'])
+						{
+							foreach ($variable['variables'] as $sub_variable) {
+								$this->template_variable_model->create(array(
+									'template_id' => 0,
+									'template_variable_id' => $var->id,
+									'type'  => isset($sub_variable['type']) ? $sub_variable['type'] : 'string',
+									'name'  => $sub_variable['name'],
+									'label' => isset($sub_variable['label']) ? $sub_variable['label'] : $sub_variable['name'],
+									'value' => isset($sub_variable['default']) ? $sub_variable['default'] : '',
+									'options' => isset($sub_variable['options']) ? serialize($sub_variable['options']) : ''
+								));
+							}
+						}
+					}
+					else
+					{
+						//update variable
+						$var = $this->template_variable_model->first(array('name' => $variable['name'], 'template_variable_id' => null, 'template_id' => 0));
+						$var->type    = $variable['type'];
+						$var->label   = isset($variable['label']) ? $variable['label'] : $variable['name'];
+						$var->value   = isset($variable['default']) ? $variable['default'] : '';
+						$var->options = isset($variable['options']) ? serialize($variable['options']) : '';
+						$var->save();
+						
+						if ($variable['type'] == 'array' && $variable['variables'])
+						{
+							foreach ($variable['variables'] as $sub_variable) {
+								
+								/*$this->template_variable_model->create(array(
+									'template_id' => $this->id,
+									'template_variable_id' => $var->id,
+									'type'  => isset($sub_variable['type']) ? $sub_variable['type'] : 'string',
+									'name'  => $sub_variable['name'],
+									'label' => isset($sub_variable['label']) ? $sub_variable['label'] : $sub_variable['name'],
+									'value' => isset($sub_variable['default']) ? $sub_variable['default'] : '',
+									'options' => isset($sub_variable['options']) ? serialize($sub_variable['options']) : ''
+								));*/
+							}
+						}
+					}
+					$variables[] = $variable['name'];
+				}
+			}
+			
+			//delete any variables that have been removed
+			foreach ($this->template_variable_model->all(array('template_id' => 0)) as $tv) {
 				if (!in_array($tv->name, $variables) && !$tv->template_variable_id) {
 					$tv->destroy();
 				}
